+++ example/android/local.properties
@@ -1,3 +1,3 @@
-flutter.sdk=/home/shivansh/snap/flutter/common/flutter
-sdk.dir=/home/shivansh/Android/Sdk
+flutter.sdk=C:\\Flutter
+sdk.dir=C:\\Users\\FrankBuijze\\AppData\\Local\\Android\\sdk
 flutter.buildMode=debug
