+++ example/lib/typed_examples/streaming.dart
