+++ example/lib/typed_examples/text_room.dart
