+++ lib/interfaces/Streaming/Events/streaming_plugin_stopping_event.dart
