+++ lib/interfaces/audio_bridge/events/audio_bridge_configured_event.dart
