+++ lib/interfaces/audio_bridge/events/audio_bridge_joined_event.dart
