+++ lib/interfaces/typed_event.dart
