+++ lib/interfaces/video_call/events/video_call_accepted_event.dart
