+++ lib/interfaces/video_call/events/video_call_registered_event.dart
