+++ lib/interfaces/video_room/events/video_room_configured.dart
