+++ lib/interfaces/video_room/events/video_room_joined_event.dart
