+++ lib/interfaces/video_room/events/video_room_leaving_event.dart
