+++ lib/interfaces/video_room/events/video_room_new_publisher_event.dart
