+++ lib/janus_client.dart
@@ -13,85 +13,85 @@ import 'package:flutter_webrtc/flutter_webrtc.dart';
 import 'package:http/http.dart' as http;
 import 'package:flutter/services.dart';
 
-part './JanusSession.dart';
+part 'janus_session.dart';
 
-part './JanusTransport.dart';
+part 'janus_transport.dart';
 
-part './JanusPlugin.dart';
+part 'janus_plugin.dart';
 
 part './utils.dart';
 
-part './WrapperPlugins/JanusVideoCallPlugin.dart';
+part './wrapper_plugins/janus_video_call_plugin.dart';
 
-part './WrapperPlugins/JanusSipPlugin.dart';
+part './wrapper_plugins/janus_sip_plugin.dart';
 
-part './WrapperPlugins/JanusVideoRoomPlugin.dart';
+part './wrapper_plugins/janus_video_room_plugin.dart';
 
-part 'interfaces/TypedEvent.dart';
+part 'interfaces/typed_event.dart';
 
-part './WrapperPlugins/JanusAudioBridgePlugin.dart';
+part './wrapper_plugins/janus_audio_bridge_plugin.dart';
 
-part './WrapperPlugins/JanusStreamingPlugin.dart';
+part './wrapper_plugins/janus_streaming_plugin.dart';
 
-part './WrapperPlugins/JanusTextRoomPlugin.dart';
+part './wrapper_plugins/janus_text_room_plugin.dart';
 
-part './WrapperPlugins/JanusEchoTestPlugin.dart';
+part './wrapper_plugins/janus_echo_test_plugin.dart';
 
-part './interfaces/VideoRoom/video_room_list_response.dart';
+part './interfaces/video_room/video_room_list_response.dart';
 
-part './interfaces/VideoRoom/video_room_list_participants_response.dart';
+part './interfaces/video_room/video_room_list_participants_response.dart';
 
-part './interfaces/VideoRoom/Events/video_room_atached_event.dart';
+part './interfaces/video_room/Events/video_room_atached_event.dart';
 
-part './interfaces/VideoRoom/Events/video_room_configured.dart';
+part './interfaces/video_room/Events/video_room_configured.dart';
 
-part './interfaces/VideoRoom/Events/video_room_joined_event.dart';
+part './interfaces/video_room/Events/video_room_joined_event.dart';
 
-part './interfaces/VideoRoom/Events/video_room_leaving_event.dart';
+part './interfaces/video_room/Events/video_room_leaving_event.dart';
 
-part './interfaces/VideoRoom/Events/video_room_new_publisher_event.dart';
+part './interfaces/video_room/Events/video_room_new_publisher_event.dart';
 
-part './interfaces/VideoRoom/Events/video_room_event.dart';
+part './interfaces/video_room/Events/video_room_event.dart';
 
-part './interfaces/VideoCall/Events/video_call_accepted_event.dart';
+part './interfaces/video_call/Events/video_call_accepted_event.dart';
 
-part './interfaces/VideoCall/Events/video_call_calling_event.dart';
+part './interfaces/video_call/Events/video_call_calling_event.dart';
 
-part './interfaces/VideoCall/Events/video_call_event.dart';
+part './interfaces/video_call/Events/video_call_event.dart';
 
-part './interfaces/VideoCall/Events/video_call_hangup_event.dart';
+part './interfaces/video_call/Events/video_call_hangup_event.dart';
 
-part './interfaces/VideoCall/Events/video_call_incoming_call_event.dart';
+part './interfaces/video_call/Events/video_call_incoming_call_event.dart';
 
-part './interfaces/VideoCall/Events/video_call_registered_event.dart';
+part './interfaces/video_call/Events/video_call_registered_event.dart';
 
-part './interfaces/Streaming/create_media_item.dart';
+part './interfaces/streaming/create_media_item.dart';
 
-part './interfaces/Streaming/streaming_mount.dart';
+part './interfaces/streaming/streaming_mount.dart';
 
-part './interfaces/Streaming/streaming_mount_edited.dart';
+part './interfaces/streaming/streaming_mount_edited.dart';
 
-part './interfaces/Streaming/streaming_mount_point.dart';
+part './interfaces/streaming/streaming_mount_point.dart';
 
-part './interfaces/Streaming/streaming_mount_point_info.dart';
+part './interfaces/streaming/streaming_mount_point_info.dart';
 
-part './interfaces/Streaming/Events/StreamingPluginPreparingEvent.dart';
+part 'interfaces/streaming/Events/streaming_plugin_preparing_event.dart';
 
-part './interfaces/Streaming/Events/StreamingPluginStoppingEvent.dart';
+part 'interfaces/streaming/Events/streaming_plugin_stopping_event.dart';
 
-part './interfaces/AudioBridge/audio_room_created_response.dart';
+part './interfaces/audio_bridge/audio_room_created_response.dart';
 
-part './interfaces/AudioBridge/rtp_forward_stopped.dart';
+part './interfaces/audio_bridge/rtp_forward_stopped.dart';
 
-part './interfaces/AudioBridge/rtp_forwarder_created.dart';
+part './interfaces/audio_bridge/rtp_forwarder_created.dart';
 
-part './interfaces/AudioBridge/Events/audio_bridge_configured_event.dart';
+part './interfaces/audio_bridge/Events/audio_bridge_configured_event.dart';
 
-part './interfaces/AudioBridge/Events/audio_bridge_event.dart';
+part './interfaces/audio_bridge/Events/audio_bridge_event.dart';
 
-part './interfaces/AudioBridge/Events/audio_bridge_joined_event.dart';
+part './interfaces/audio_bridge/Events/audio_bridge_joined_event.dart';
 
-part './interfaces/AudioBridge/Events/audio_bridge_leaving_event.dart';
+part './interfaces/audio_bridge/Events/audio_bridge_leaving_event.dart';
 
 class JanusClient {
   late JanusTransport _transport;
